Add bike repair locations to open data
We've collected several datasets related to sustainability on campus. We would like these datasets to make their way into the open data API, some of which for eventual inclusion on the campus map.

The data were collected by Sedona Cluett (sustainability.coop@uwaterloo.ca), the sustainability co-op for this term working for Mat Thijssen (mthijssen@uwaterloo.ca), UW's sustainability coordinator. Please contact one of us if there are issues with this request or any changes that need to be made to the dataset before it can be merged into the master. Thanks! 

James (jmccarth@uwaterloo.ca)